+++ Developer/AppFrameworks/UIKit/YJTableView/YJTableView/YJFirstViewController.m
@@ -76,7 +76,7 @@
 
 #pragma mark YJUIScrollViewManagerDelegate
 - (void)scrollViewManager:(YJUIScrollViewManager *)manager didVerticalScroll:(YJUIScrollViewScroll)scroll {
-    NSLog(@"%@ -- %d", NSStringFromSelector(_cmd), scroll);
+    NSLog(@"%@ -- %ld", NSStringFromSelector(_cmd), (long)scroll);
     if (scroll == YJUIScrollViewScrollEndBottom) { // 分页加载
         NSLog(@"%@分页加载", NSStringFromSelector(_cmd));
         [self initTestData];
